Dev ID selection for widget and refactoring the class structure (#301)
* Changing state to json string

* Implementing data changes on Assigned to me widget and adding account selection in Issues and PR templates

* Convertiong Assigned to Me and Mentioned Me widgets to one common base class for them to work with JSON ConfigurationData strings.

* Adding common get devids code

* Adding multiaccount methods to category widgets

* Refactoring Review widget to work alongside Assigned to me and Mentioned in

* Refactoring Issues and PR widget into a common class

* More refactoring and name changes

* Comment

* Changing UserName method.

* nit: devId -> developerId

* Removing base Reset Info method and fixing bug on reseting Review Widget

* Fixed some of the flow. CanPin and Pinned not needed anymore. Changed to Saved.

* Fixing customization request

* nit

* Fix Saved on Cancel

---------
