+++ src/GitHubExtension/Widgets/GitHubIssuesWidget.cs
@@ -10,22 +10,11 @@ using Microsoft.Windows.Widgets.Providers;
 using Octokit;
 
 namespace GitHubExtension.Widgets;
-internal class GitHubIssuesWidget : GitHubWidget
+internal class GitHubIssuesWidget : GitHubRepositoryWidget
 {
     private readonly string issuesIconData = IconLoader.GetIconAsBase64("issues.png");
 
     protected static readonly new string Name = nameof(GitHubIssuesWidget);
-
-    public GitHubIssuesWidget()
-    : base()
-    {
-        GitHubDataManager.OnUpdate += DataManagerUpdateHandler;
-    }
-
-    ~GitHubIssuesWidget()
-    {
-        GitHubDataManager.OnUpdate -= DataManagerUpdateHandler;
-    }
 
     public override void DeleteWidget(string widgetId, string customState)
     {
@@ -192,19 +181,7 @@ internal class GitHubIssuesWidget : GitHubWidget
         };
     }
 
-    public override string GetData(WidgetPageState page)
-    {
-        return page switch
-        {
-            WidgetPageState.SignIn => GetSignIn(),
-            WidgetPageState.Configure => GetConfiguration(RepositoryUrl),
-            WidgetPageState.Content => ContentData,
-            WidgetPageState.Loading => EmptyJson,
-            _ => throw new NotImplementedException(Page.GetType().Name),
-        };
-    }
-
-    private void DataManagerUpdateHandler(object? source, DataManagerUpdateEventArgs e)
+    protected override void DataManagerUpdateHandler(object? source, DataManagerUpdateEventArgs e)
     {
         Log.Logger()?.ReportDebug(Name, ShortId, $"Data Update Event: Kind={e.Kind} Info={e.Description} Context={string.Join(",", e.Context)}");
 
@@ -224,11 +201,5 @@ internal class GitHubIssuesWidget : GitHubWidget
                 UpdateActivityState();
             }
         }
-    }
-
-    public override void OnCustomizationRequested(WidgetCustomizationRequestedArgs customizationRequestedArgs)
-    {
-        SavedRepositoryUrl = RepositoryUrl;
-        SetConfigure();
     }
 }
