+++ src/GitHubExtension/Widgets/Templates/GitHubAssignedConfigurationTemplate.json
@@ -1,56 +1,69 @@
 {
-    "type": "AdaptiveCard",
-    "$schema": "http://adaptivecards.io/schemas/adaptive-card.json",
-    "version": "1.5",
-    "body": [
+  "type": "AdaptiveCard",
+  "$schema": "http://adaptivecards.io/schemas/adaptive-card.json",
+  "version": "1.5",
+  "body": [
+    {
+      "type": "Input.ChoiceSet",
+      "id": "account",
+      "placeholder": "%Widget_Template/ChooseAccountPlaceholder%",
+      "value": "${selectedDevId}",
+      "choices": [
         {
-            "type": "ColumnSet",
-            "columns": [
+          "$data": "${accounts}",
+          "title": "${devid}",
+          "value": "${devid}"
+        }
+      ]
+    },
+    {
+      "type": "ColumnSet",
+      "columns": [
+        {
+          "type": "Column",
+          "width": "stretch",
+          "items": [
+            {
+              "type": "Input.ChoiceSet",
+              "id": "showCategory",
+              "style": "compact",
+              "isMultiSelect": false,
+              "value": "${showCategory}",
+              "choices": [
                 {
-                    "type": "Column",
-                    "width": "stretch",
-                    "items": [
-                        {
-                            "type": "Input.ChoiceSet",
-                            "id": "ShowCategory",
-                            "style": "compact",
-                            "isMultiSelect": false,
-                            "value": "${showCategory}",
-                            "choices": [
-                                {
-                                    "title": "%Widget_Template/Issues%",
-                                    "value": "Issues"
-                                },
-                                {
-                                    "title": "%Widget_Template/PullRequests%",
-                                    "value": "PullRequests"
-                                },
-                                {
-                                    "title": "%Widget_Template/IssuesAndPullRequests%",
-                                    "value": "IssuesAndPullRequests"
-                                }
-                            ]
-                        }
-                    ]
+                  "title": "%Widget_Template/Issues%",
+                  "value": "Issues"
                 },
                 {
-                    "type": "Column",
-                    "width": "auto",
-                    "items": [
-                        {
-                            "type": "ActionSet",
-                            "actions": [
-                                {
-                                    "type": "Action.Execute",
-                                    "title": "%Widget_Template_Button/Submit%",
-                                    "verb": "Submit",
-                                    "associatedInputs": "auto"
-                                }
-                            ]
-                        }
-                    ]
+                  "title": "%Widget_Template/PullRequests%",
+                  "value": "PullRequests"
+                },
+                {
+                  "title": "%Widget_Template/IssuesAndPullRequests%",
+                  "value": "IssuesAndPullRequests"
                 }
-            ]
+              ]
+            }
+          ]
+        },
+        {
+          "type": "Column",
+          "width": "auto",
+          "items": [
+            {
+              "type": "ActionSet",
+              "actions": [
+                {
+                  "type": "Action.Execute",
+                  "title": "%Widget_Template_Button/Submit%",
+                  "verb": "Submit",
+                  "associatedInputs": "auto"
+                }
+              ]
+            }
+          ]
         }
-    ]
+      ]
+    }
+  ]
 }
