+++ src/GitHubExtension/Widgets/Templates/GitHubReviewTemplate.json
@@ -22,7 +22,7 @@
           "items": [
             {
               "type": "TextBlock",
-              "text": "@${referredName}",
+              "text": "${userName}",
               "size": "Large",
               "style": "Heading",
               "wrap": true
